video-pipeline/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0

# Video Processing
moviepy==1.0.3
imageio==2.31.6
imageio-ffmpeg==0.4.9
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.4
scipy==1.11.4

# Subtitle Processing
pysrt==1.1.2
python-ffmpeg==2.0.12
ffmpeg-python==0.2.0

# Image Processing
requests==2.31.0
beautifulsoup4==4.12.2
unsplash-python==1.0.0a2

# Text Processing
textwrap3==0.9.2
fonttools==4.44.0

# Utilities
structlog==23.2.0
colorama==0.4.6
rich==13.7.0
typer==0.9.0
prometheus-client==0.19.0
jinja2==3.1.2
pyyaml==6.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1